+++ nodalia-cards.manifest.js
@@ -1,10 +1,10 @@
 export default {
-  "pkgVersion": "1.1.1-alpha.8",
-  "contentSha256_12": "6e98f7eaa999",
+  "pkgVersion": "1.1.1-alpha.9",
+  "contentSha256_12": "537f97606aba",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
-  "hacsFile": "nodalia-cards-1.1.1-alpha.8.js"
+  "hacsFile": "nodalia-cards-1.1.1-alpha.9.js"
 };
-export const pkgVersion = "1.1.1-alpha.8";
-export const contentSha256_12 = "6e98f7eaa999";
+export const pkgVersion = "1.1.1-alpha.9";
+export const contentSha256_12 = "537f97606aba";
 export const file = "nodalia-cards.bundle.js";